renamed sequencer track types

+++ src/framework/audio/devtools/audioenginedevtools.cpp
@@ -189,7 +189,7 @@ void AudioEngineDevTools::stop()
     sequencer()->stop();
 }
 
-void AudioEngineDevTools::setLoop(unsigned int from, unsigned int to)
+void AudioEngineDevTools::setLoop(uint64_t from, uint64_t to)
 {
     sequencer()->setLoop(from, to);
 }
